+++ src/pynwb/ndx_pose/io/pose.py
@@ -9,6 +9,7 @@ from ..pose import PoseEstimation, PoseEstimationSeries
 class PoseEstimationSeriesMap(TimeSeriesMap):
 
     def __init__(self, spec):
+        """Map attribute spec "definition" to Python instance attribute "confidence_definition"."""
         super().__init__(spec)
         confidence_spec = self.spec.get_dataset('confidence')
         self.map_spec('confidence_definition', confidence_spec.get_attribute('definition'))
@@ -18,6 +19,7 @@ class PoseEstimationSeriesMap(TimeSeriesMap):
 class PoseEstimationMap(NWBContainerMapper):
 
     def __init__(self, spec):
+        """Map attribute spec "version" to Python instance attribute "source_software_version"."""
         super().__init__(spec)
         source_software_spec = self.spec.get_dataset('source_software')
         self.map_spec('source_software_version', source_software_spec.get_attribute('version'))